requirements.txt
aiohttp==3.7.4.post0
argcomplete==2.0.0
async-timeout==3.0.1
attrs==21.4.0
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
click==8.1.3
colorama==0.4.4
discord.py==1.7.3
distlib==0.3.4
ffmpeg==1.4
filelock==3.6.0
idna==3.3
lavalink==3.1.5
multidict==6.0.2
packaging==21.3
pipenv==2022.5.2
pipx==1.0.0
platformdirs==2.5.2
pycparser==2.21
PyNaCl==1.5.0
pyparsing==3.0.8
six==1.16.0
typing_extensions==4.2.0
userpath==1.8.0
virtualenv==20.14.1
virtualenv-clone==0.5.7
yarl==1.7.2
youtube-dl==2021.12.17